requirements.txt
click
plotly == 4.0.0
selenium
python-dateutil
dash
dash-daq
mypy
pylint